+++ lib/src/widgets/app_lock.dart
@@ -10,14 +10,21 @@ import 'package:flutter/material.dart';
 /// [Widget]. The [Object] argument is provided by the [lockScreen] calling
 /// `AppLock.of(context).didUnlock();` with an argument. [Object] can then be injected
 /// in to your `MyApp` widget (or equivalent).
+///
+/// [enabled] determines wether or not the [lockScreen] should be shown on app launch
+/// and subsequent app pauses. This can be changed later on using `AppLock.of(context).enable();`,
+/// `AppLock.of(context).disable();` or the convenience method `AppLock.of(context).setEnabled(enabled);`
+/// using a bool argument.
 class AppLock extends StatefulWidget {
   final Widget Function(Object) builder;
   final Widget lockScreen;
+  final bool enabled;
 
   const AppLock({
     Key key,
     @required this.builder,
     @required this.lockScreen,
+    this.enabled = true,
   }) : super(key: key);
 
   static _AppLockState of(BuildContext context) =>
@@ -32,19 +39,25 @@ class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
 
   bool _didUnlockForAppLaunch;
   bool _isPaused;
+  bool _enabled;
 
   @override
   void initState() {
     WidgetsBinding.instance.addObserver(this);
 
-    this._didUnlockForAppLaunch = false;
+    this._didUnlockForAppLaunch = !this.widget.enabled;
     this._isPaused = false;
+    this._enabled = this.widget.enabled;
 
     super.initState();
   }
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
+    if (!this._enabled) {
+      return;
+    }
+
     if (state == AppLifecycleState.paused &&
         (!this._isPaused && this._didUnlockForAppLaunch)) {
       this._showLockScreen();
@@ -63,7 +76,7 @@ class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: this._lockScreen,
+      home: this.widget.enabled ? this._lockScreen : this.widget.builder(null),
       navigatorKey: _navigatorKey,
       routes: {
         '/lock-screen': (context) => this._lockScreen,
@@ -80,12 +93,48 @@ class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
     );
   }
 
+  /// Causes `AppLock` to either pop the [lockScreen] if the app is already running
+  /// or instantiates widget returned from the [builder] method if the app is cold
+  /// launched.
+  ///
+  /// [args] is an optional argument which will get passed to the [builder] method
+  /// when built. Use this when you want to inject objects created from the
+  /// [lockScreen] in to the rest of your app so you can better guarantee that some
+  /// objects, services or databases are already instantiated before using them.
   void didUnlock([Object args]) {
     if (this._didUnlockForAppLaunch) {
       this._didUnlockOnAppPaused();
     } else {
       this._didUnlockOnAppLaunch(args);
     }
+  }
+
+  /// Makes sure that [AppLock] shows the [lockScreen] on subsequent app pauses if
+  /// [enabled] is true of makes sure it isn't shown on subsequent app pauses if
+  /// [enabled] is false.
+  ///
+  /// This is a convenience method for calling the [enable] or [disable] method based
+  /// on [enabled].
+  void setEnabled(bool enabled) {
+    if (enabled) {
+      this.enable();
+    } else {
+      this.disable();
+    }
+  }
+
+  /// Makes sure that [AppLock] shows the [lockScreen] on subsequent app pauses.
+  void enable() {
+    setState(() {
+      this._enabled = true;
+    });
+  }
+
+  /// Makes sure that [AppLock] doesn't show the [lockScreen] on subsequent app pauses.
+  void disable() {
+    setState(() {
+      this._enabled = false;
+    });
   }
 
   void _didUnlockOnAppLaunch(Object args) {